fix: parse_test_tag: update regexp to allow underscores
The spec initially restricted the test ID modifier to alphanumeric
characters, but this doesn't work well in the context of
metworxinspector, where test_config_pkgs() tacks on package names as
the modifier element.  To handle this, the spec has been extended to
allow underscores and metworxinspector now sanitizes non-alphanumeric
characters to underscores.