dev: read_requirements_gsheet: use columns that match upstream sheet
Originally the sheet used names with spaces, so for convenience
read_requirements_gsheet mapped the names to snake case.  Now that the
requirements sheet consistently uses CamelCase, use the same column
names so that people don't need to keep the "sheet to internal data
frame" column mapping in their head.

The *_col parameters of read_requirements_gsheet() could be removed
altogether if we want to rely on the input sheets using the pre-chosen
names, but since we already have the flexibility, leave the parameters
for now.