dev: support reading stories
As with the requirement sheet, it might not be worth keeping the *_col
parameters for mapping column names to the ones we expect, but add it
for now since read_requirements_gsheet() still supports it.

The sotries still need to be joined with the requirements, which we'll
probably want a dedicated helper for.