fix: tweak capitalization in table headers
Note that Aing didn't explicitly request the changes in
write_requirements()/format_spec(), but those are consistent with
what's suggested for the other tables.

Also note that the current result is still inconsistent.  For example,
"passed" and "failed" aren't capitalized.  And "Story" in "User Story"
of the traceability matrix table is capitalized, which is consistent
with its appearance in the requirements section headers but
inconsistent with the lowercase "name" in "Test name".