feat: write_traceability_matrix: put test IDs in same cell
write_traceability_matrix() generates a table with user stories and
test IDs.  The same user story can span multiple rows, with the test
ID changing.  In this case, the repeated user story is left blank to
make the grouping stand out.

A downside of this approach is that the story is spread across
multiple rows and can't be easily highlighted by simple borders or
alternating colors (which could be done via a reference docx style).

To make this possible, instead put all the test IDs into the same
cell.  This has the disadvantage that the row needs to be manually
resized after the fact so that each test ID is on its own
line.  **But** this is already the case: cells with a single ID are
being broken across multiple lines, and I haven't found a way to fix
that (including trying the pander package).  So, if we're going to
need to manually resize the column anyway, we might as well use a
single row so that the groups can be easily styled.