fix: write_requirements: restore summary of requirements
The requirements document is supposed to include a "Summary" header
followed by a list of requirements.  The list of requirements was lost
in 13657f2 (feat: make requirements optional for specs, 2021-08-04).
The problem is a vectorized ifelse(SCALAR, "", VECTOR) call that leads
to dropping all of VECTOR except for the first item ("**Summary**\n").

Use a plain if/else instead.