fix: write_traceability_matrix: show only first StoryDescription line
write_traceability_matrix() dumps the entire StoryDescription to the
"User Story" column.  This looks okay for the stories coming from our
current Google sheet, but untamed input can break the table structure.
Stories generated from GitHub issues, for example, contain a
requirements list in the body and this breaks the matrix table output.

Filter to the first line, as anything beyond that is likely too
verbose anyway.  If we later run into other input the breaks the
table, we may need to consider sanitizing the story text.