fix: create_validation_docs: drop orphan IDs in testing doc
Unlike the other write_* functions, write_validation_testing() uses
the test data frame directly rather than using the joined data frame
that links the stories and requirements.  Because of this, a test ID
that's not mentioned in any of the other docs can sneak into the
testing doc.

To prevent confusing entries in the testing doc, warn about these rows
and drop them.