fix: get_reference_docx: return absolute path
The documentation of get_reference_docx() claims to return a full
path, but it actually returns a relative path.  Ignoring the doc/code
disagreement, at first glance that should be okay because the relative
path still points to the right place.

The problem is that rmarkdown::render() sets the current working
directory to the input file's directory.  render() passes the
reference docx path as is to `pandoc`, so the relative path will end
up pointing to the wrong place if the input file isn't in the current
directory.  In the context of mrgvalidate, this happens when
`output_dir` is set to something other than the current working
directory (the default value).

Sidestep the issue by making get_reference_docx() follow its already
documented behavior: return an absolute path.