test: adds a failing test case that will be fixed in the next commit.
This commit adds a row to specs.RDS that attaches a new VSC-R003 to requirement to the VSC-S001 story. VSC-R003 is satisfied only by the test MXI-VSC-001, which is the same test that satisfies (previously existing) VSC-R003.

The bug, to be fixed in the next commit, is that when multiple requirements are satisfied by the same test(s) and attached to the same story, only the first of them is included in the Requirements Specification document.