fix: Addresses issue #89 and fixes failing test case added in 48fc0c2
Specifically, the bug is that when multiple requirements are satisfied by the same test(s) and attached to the same story, only the first of them is included in the Requirements Specification document.

This was caused by an errant distinct() call in make_requirements() that has essentially been moved to format_req_tests().